requirements.txt
Django==1.10.5
haikunator==2.1.0
MySQL-python==1.2.5
python-ldap==2.4.32
python-owasp-zap-v2.4==0.0.8
requests==2.13.0
